Start rewrite of code-base using Laravel 5.3
This constitutes the beginnnings of a from-scratch rewrite of the code
base using Laravel 5. The intent is to take advantage of an opportunity
to re-structure the application to make some of the longest standing
feature requests easier to implement such as file uploading, tagging,
and a more complete API that can be consumed by third party tools.

This commit includes the adusted database structure, and a functional
implementation of the previous API but is not a complete application.