Expose UUID as the route key and id for Mod and Modpack
As a carryover from the v0.7 API, the Mod and Modpack models and API endpoints were using an arbitrary slug as the route key; while this is easier for humans to read, it only complicates the api (which is designed to be consumed by machines). The slug will remain for backwards compatibility with the v0.7 API, but should probably be considered for deprecation in the future.