Shift tests
Laravel 5.4 made significant changes to testing and extracted many of the methods being used in the original tests into a support package. In the interest of keeping the application lean I’ve opted to re-create the functionality of the tests from scratch instead of directly porting the tests that existed. This means there WILL be a change in test coverage (probably a reduction) and I ended up doing some level of refactoring and bug fixing as part of the development.