Simplify privacy icons and give icons some breathing room
Setup some icon aliases so we don’t have to have a big if, else if block in the view. Might be good to expand this to other site icons too, makes keeping them consistent easier and if they ever need to be changed its only one location, not hundreds.

Also, removed the use of bulma's .icon class and replaced it with the font awesome fixed with .fa-fw class on the icon. The end result is subtle but significant, .icon locks the icon to 21px wide where .fa-fw is 28px (both at normal font size) which gives just a few pixels of space around the icon to make it feel more naturally part of the element. It may even serve the same ultimate purpose of the .icon class, which is to fix the space on the page even before the icon is loaded, although I have no good way to test that.